Extended .arclint
Summary:
Added more linters to the .arclint setup.

Note that it will hate on you the first time you edit a file with Tabs instead
of spaces (there are a few of those hanging around). Make sure to update
the editor to auto-correct those.

Test Plan: Try `arc lint --everything` and cry.

Reviewers: #kvalobs, vegardb

Reviewed By: #kvalobs, vegardb

Subscribers: vegardb

Differential Revision: https://phab.met.no/D876