Ensuring data is published on Kafka
Summary:
Second cut on implementing a solution to this. This one is a little more elegant, and hopefully also works correctly.

Basically:
- Loop over the processing.
- Set processingCompleted to false before attempting to contact Kafka.
- If Kafka returns without an error, it sets processingCompleted to true.
- After Kafka, we check processingCompleted. If true,  we mark the station info done, and reset the processTimer delay to 1. If false, we  sleep for `processTimer` seconds,  and increase processTimer.

Note, this should happen for both cases where Kafka fails (for whatever reason) and cases where the process does not get any stationinfo to process from the database.

The delay timer is currently set to peak at 5 minutes (300 seconds). It would probably make sense for this to be configurable eventually.

Test Plan: Build and run this version of kvQabased against a running database/kafka instance. Verify that it processes data correctly. Check the error cases.

Reviewers: borgem, vegardb

Reviewed By: vegardb

Differential Revision: https://phab.met.no/D1549